pyproject.toml
[project]
name = "dataflow-doc-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asttokens==2.0.5",
    "backcall==0.2.0",
    "colorama==0.4.4",
    "cryptography>=44.0.2",
    "decorator==5.1.1",
    "executing==0.8.3",
    "flake8==4.0.1",
    "ipython==8.3.0",
    "jedi==0.18.1",
    "jinja2==3.1.2",
    "jsonpickle==2.2.0",
    "loguru==0.7.2",
    "markupsafe==2.1.1",
    "matplotlib-inline==0.1.3",
    "mccabe==0.6.1",
    "neo4j-driver==5.18.0",
    "networkx==3.1",
    "parso==0.8.3",
    "pickleshare==0.7.5",
    "prompt-toolkit==3.0.29",
    "pure-eval==0.2.2",
    "pycodestyle==2.8.0",
    "pyflakes==2.4.0",
    "pygments==2.12.0",
    "pyvis==0.3.2",
    "pyyaml==6.0.1",
    "six==1.16.0",
    "stack-data==0.2.0",
    "traitlets==5.2.0",
    "wcwidth==0.2.5",
]
